MAKE EVERYTHING GOOD

+++ commands/serverstats.js
@@ -1,16 +1,13 @@
 const Discord = require("discord.js");
 const db = require("quick.db");
 const serverstats = new db.table("ServerStats");
+require('dotenv').config({ path: '../.env' })
 
 exports.run = async (client, message, args, tools) => {
-  if (!message.member.hasPermission("MANAGE_GUILD"))
-    return message.channel.send(
-      `:x: You need **MANAGE_GUILD** permission to use this command.`
-    );
+  if (!message.member.permissions.has(process.env.SETUP_NEEDED_PERMISSION))
+    return message.channel.send(process.env.LANGUAGE_NO_PERMISSION);
   if (!args[0])
-    return message.channel.send(
-      ":x: Invalid parameters. Correct usage: `!serverstats enable` | `!serverstats disable`."
-    );
+    return message.channel.send(process.env.LANGUAGE_INVALID_PARAMETERS);
   if (args[0] === "enable") {
     let totusers = await serverstats.fetch(`Stats_${message.guild.id}`, {
       target: ".totusers"
@@ -22,52 +19,49 @@ exports.run = async (client, message, args, tools) => {
       target: ".botcount"
     });
     if (totusers !== null || membcount !== null || botcount !== null)
-      return message.channel.send(
-        `:x: Server stats are already enabled for this server.`
-      );
-    if (!message.guild.me.hasPermission(`MANAGE_CHANNELS`))
-      return message.channel.send(
-        `:x: I don't have **MANAGE_CHANNELS** permission.`
-      );
-    const totalsize = message.guild.memberCount;
-    const botsize = message.guild.members.filter(m => m.user.bot).size;
-    const humansize = totalsize - botsize;
+      return message.channel.send(process.env.LANGUAGE_ALREADY_ENABLED);
+    if (!message.guild.me.permissions.has(`MANAGE_CHANNELS`))
+      return message.channel.send(process.env.LANGUAGE_BOT_MISSING_PERMISSION);
     message.guild
-      .createChannel("📈Server Statistics📈", "category", [
-        {
-          id: message.guild.id,
-          deny: ["CONNECT"]
-        }
-      ])
+      .channels.create(process.env.LANGUAGE_SERVER_STATS, {
+        type: 'GUILD_CATEGORY', permissionOverwrites: [
+          {
+            id: message.guild.id,
+            deny: ["CONNECT"]
+          }]
+      })
       .then(channel => {
         channel.setPosition(0);
         message.guild
-          .createChannel("Total Users : " + totalsize, "voice", [
-            {
-              id: message.guild.id,
-              deny: ["CONNECT"]
-            }
-          ])
+          .channels.create(process.env.LANGUAGE_TOTAL_USERS + message.guild.memberCount, {
+            type: 'GUILD_VOICE', permissionOverwrites: [
+              {
+                id: message.guild.id,
+                deny: ["CONNECT"]
+              }]
+          })
           .then(channel1 => {
             channel1.setParent(channel.id);
             let x = channel1.id;
             message.guild
-              .createChannel("Human Users  : " + humansize, "voice", [
-                {
-                  id: message.guild.id,
-                  deny: ["CONNECT"]
-                }
-              ])
+              .channels.create(process.env.LANGUAGE_HUMAN_USERS + message.guild.members.cache.filter(member => !member.user.bot).size, {
+                type: 'GUILD_VOICE', permissionOverwrites: [
+                  {
+                    id: message.guild.id,
+                    deny: ["CONNECT"]
+                  }]
+              })
               .then(channel2 => {
                 channel2.setParent(channel.id);
                 let y = channel2.id;
                 message.guild
-                  .createChannel("Bot Users : " + botsize, "voice", [
-                    {
-                      id: message.guild.id,
-                      deny: ["CONNECT"]
-                    }
-                  ])
+                  .channels.create(process.env.LANGUAGE_BOT_USERS + message.guild.members.cache.filter(member => member.user.bot).size, {
+                    type: 'GUILD_VOICE', permissionOverwrites: [
+                      {
+                        id: message.guild.id,
+                        deny: ["CONNECT"]
+                      }]
+                  })
                   .then(async channel3 => {
                     channel3.setParent(channel.id);
                     let z = channel3.id;
@@ -82,9 +76,7 @@ exports.run = async (client, message, args, tools) => {
               });
           });
       });
-    message.channel.send(
-      `:white_check_mark: Serverstats enabled for this server.`
-    );
+    message.channel.send(process.env.LANGUAGE_ENABLED_SERVERSTATS);
   } else if (args[0] === "disable") {
     let totusers = await serverstats.fetch(`Stats_${message.guild.id}`, {
       target: ".totusers"
@@ -99,17 +91,21 @@ exports.run = async (client, message, args, tools) => {
       target: ".categid"
     });
     if (totusers === null || membcount === null || botcount === null)
-      return message.channel.send(
-        `:x: Serverstats for this server is not enabled.`
-      );
-    client.channels.cache.get(totusers).delete();
-    client.channels.cache.get(membcount).delete();
-    client.channels.cache.get(botcount).delete();
-    client.channels.cache.get(categ).delete();
+      return message.channel.send(process.env.LANGUAGE_DISABLED_SERVERSTATS_CURRENTLY);
 
+    if (client.channels.cache.get(totusers) != undefined) {
+      client.channels.cache.get(totusers).delete();
+    }
+    if (client.channels.cache.get(membcount) != undefined) {
+      client.channels.cache.get(membcount).delete();
+    }
+    if (client.channels.cache.get(botcount) != undefined) {
+      client.channels.cache.get(botcount).delete();
+    }
+    if (client.channels.cache.get(categ) != undefined) {
+      client.channels.cache.get(categ).delete();
+    }
     serverstats.delete(`Stats_${message.guild.id}`);
-    message.channel.send(
-      `:white_check_mark: Serverstats disabled for this server.`
-    );
+    message.channel.send(process.env.LANGUAGE_DISABLED_SERVERSTATS);
   }
 };
